refactor(waitlist): quebrar WaitlistSection em subcomponentes
Extrai WaitlistCopyPanel, WaitlistFormFields, WaitlistFormMessages e
WaitlistSubmitButton para components/landings/lollapalooza/waitlist/,
resolvendo o warning react-doctor/no-giant-component. Refactor puro,
sem mudança de comportamento — ids, roles e a estrutura de seções
permanecem os mesmos.

Refs #1095
